Refactored into separate class libraries. Added Loconet class library
