+++ ThrottleX.sln
@@ -8,14 +8,18 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "projectFiles", "projectFiles", "{E0BFA130-CA27-4D45-B8BB-D8042EC11CF6}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
+		.gitignore = .gitignore
 		global.json = global.json
 		README.md = README.md
-		.gitignore = .gitignore
 	EndProjectSection
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{EF36F24A-097A-4EEA-A400-C95DD50F195F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ThrottleX.Core", "src\ThrottleX.Core\ThrottleX.Core.csproj", "{EF8DC1AA-3116-4203-A149-BC8CA4220909}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ThrottleX.Core", "src\ThrottleX.Core\ThrottleX.Core.csproj", "{EF8DC1AA-3116-4203-A149-BC8CA4220909}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Loconet", "src\Loconet\Loconet.csproj", "{F8B5BAF8-A269-4C23-AD95-E64DF78A3672}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WiThrottle", "src\WiThrottle\WiThrottle.csproj", "{6122FA7A-DDED-4D2E-841A-324B51CD8A36}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +31,14 @@ Global
 		{EF8DC1AA-3116-4203-A149-BC8CA4220909}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EF8DC1AA-3116-4203-A149-BC8CA4220909}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EF8DC1AA-3116-4203-A149-BC8CA4220909}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F8B5BAF8-A269-4C23-AD95-E64DF78A3672}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F8B5BAF8-A269-4C23-AD95-E64DF78A3672}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F8B5BAF8-A269-4C23-AD95-E64DF78A3672}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F8B5BAF8-A269-4C23-AD95-E64DF78A3672}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6122FA7A-DDED-4D2E-841A-324B51CD8A36}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6122FA7A-DDED-4D2E-841A-324B51CD8A36}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6122FA7A-DDED-4D2E-841A-324B51CD8A36}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6122FA7A-DDED-4D2E-841A-324B51CD8A36}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -34,6 +46,8 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{E0BFA130-CA27-4D45-B8BB-D8042EC11CF6} = {E6C11790-D272-43C0-AB68-50B700AAAA8C}
 		{EF8DC1AA-3116-4203-A149-BC8CA4220909} = {E6C11790-D272-43C0-AB68-50B700AAAA8C}
+		{F8B5BAF8-A269-4C23-AD95-E64DF78A3672} = {E6C11790-D272-43C0-AB68-50B700AAAA8C}
+		{6122FA7A-DDED-4D2E-841A-324B51CD8A36} = {E6C11790-D272-43C0-AB68-50B700AAAA8C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {BBFAD6BB-EAF3-45AC-9A98-551631B682CD}
